refactor(server): finish route_bridge module downsizing
Code-motion-only round-5 follow-up: delivery/ordered.rs →
ordered/ordered_send/channels; delivery/remote.rs →
remote/remote_socket/remote_route_helpers; registration/owner.rs →
owner/owner_retire/owner_update; registration/socket.rs →
socket/socket_forwarder. Function inventory unchanged; every
route_bridge module is now at or under ~400 lines.
