+++ server/crates/waddle-server/src/clustering/route_bridge/registration/mod.rs
@@ -1,6 +1,9 @@
 mod owner;
+mod owner_retire;
+mod owner_update;
 mod side_effects;
 mod socket;
+mod socket_forwarder;
 
 pub(super) use owner::owner_remote_entry_if_current;
 pub(super) use side_effects::{
