+++ server/crates/waddle-server/src/clustering/route_bridge/registration/owner.rs
@@ -1,5 +1,8 @@
 use super::super::delivery::receiver::{current_claim, user_entity};
 use super::super::*;
+use super::owner_update::remote_owner_registration_is_current;
+
+pub(in super::super) use super::owner_update::owner_remote_entry_if_current;
 
 impl OrderedRelayDeliveryBridge {
     pub(super) async fn remove_remote_owner_registration_if_current(
@@ -293,255 +296,6 @@ impl OrderedRelayDeliveryBridge {
         }
     }
 
-    pub(crate) async fn update_remote_user_resource_on_owner(
-        &self,
-        msg: RelayUpdateRemoteUserResource,
-    ) -> RelayRemoteResourceUpdateReply {
-        let Some(services) = self.services.get().cloned() else {
-            return RelayRemoteResourceUpdateReply {
-                status: RelayRemoteResourceUpdateStatus::Unavailable,
-            };
-        };
-        let registration = self
-            .remote_owner_resources
-            .lock()
-            .await
-            .get(&msg.jid)
-            .filter(|registration| {
-                registration.registration_id == msg.registration_id
-                    && registration.socket_generation == msg.socket_generation
-            })
-            .cloned();
-        let Some(registration) = registration else {
-            return RelayRemoteResourceUpdateReply {
-                status: RelayRemoteResourceUpdateStatus::StaleRegistration,
-            };
-        };
-        let actor = match services
-            .user_registry
-            .ask(waddle_xmpp::registry::GetUser {
-                bare_jid: msg.jid.to_bare(),
-            })
-            .mailbox_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-            .reply_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-            .await
-        {
-            Ok(Some(actor)) => actor,
-            Ok(None) => {
-                return RelayRemoteResourceUpdateReply {
-                    status: RelayRemoteResourceUpdateStatus::StaleRegistration,
-                };
-            }
-            Err(error) => {
-                tracing::warn!(
-                    jid = %msg.jid,
-                    %error,
-                    "clustered remote-resource state update could not resolve owner UserActor"
-                );
-                return RelayRemoteResourceUpdateReply {
-                    status: RelayRemoteResourceUpdateStatus::Unavailable,
-                };
-            }
-        };
-        let jid = msg.jid;
-        let status = match msg.update {
-            RemoteResourceStateUpdate::Presence {
-                available,
-                priority,
-                state,
-            } => {
-                match owner_remote_entry_if_current(
-                    &actor,
-                    &services.connection_registry,
-                    &jid,
-                    &registration.owner,
-                )
-                .await
-                {
-                    Ok(_) => {
-                        if apply_remote_resource_presence_to_registry(
-                            &services.connection_registry,
-                            &jid,
-                            &registration.owner,
-                            available,
-                            priority,
-                            state,
-                        ) {
-                            RelayRemoteResourceUpdateStatus::Updated
-                        } else {
-                            RelayRemoteResourceUpdateStatus::StaleRegistration
-                        }
-                    }
-                    Err(status) => status,
-                }
-            }
-            RemoteResourceStateUpdate::Carbons { enabled } => {
-                match owner_remote_entry_if_current(
-                    &actor,
-                    &services.connection_registry,
-                    &jid,
-                    &registration.owner,
-                )
-                .await
-                {
-                    Ok(entry) => {
-                        entry.carbons_enabled.store(enabled, Ordering::Relaxed);
-                        RelayRemoteResourceUpdateStatus::Updated
-                    }
-                    Err(status) => status,
-                }
-            }
-            RemoteResourceStateUpdate::RosterInterested => match owner_remote_entry_if_current(
-                &actor,
-                &services.connection_registry,
-                &jid,
-                &registration.owner,
-            )
-            .await
-            {
-                Ok(entry) => {
-                    entry.roster_interested.store(true, Ordering::Relaxed);
-                    RelayRemoteResourceUpdateStatus::Updated
-                }
-                Err(status) => status,
-            },
-            RemoteResourceStateUpdate::BlocklistInterested => {
-                match owner_remote_entry_if_current(
-                    &actor,
-                    &services.connection_registry,
-                    &jid,
-                    &registration.owner,
-                )
-                .await
-                {
-                    Ok(entry) => {
-                        entry.blocklist_interested.store(true, Ordering::Relaxed);
-                        RelayRemoteResourceUpdateStatus::Updated
-                    }
-                    Err(status) => status,
-                }
-            }
-        };
-        RelayRemoteResourceUpdateReply { status }
-    }
-
-    pub(super) async fn retire_remote_owner_registration(
-        &self,
-        services: &OrderedRelayDeliveryServices,
-        jid: &jid::FullJid,
-        registration: &RemoteOwnerRegistration,
-    ) -> bool {
-        let mut handle =
-            RelayHandle::new(registration.socket_node.clone(), self.stop_token.clone())
-                .with_ask_timeouts(self.mailbox_timeout, self.reply_timeout);
-        let detach = handle
-            .force_detach_remote_user_resource(RelayForceDetachRemoteUserResource {
-                jid: jid.clone(),
-                registration_id: registration.registration_id,
-                requester_bare_jid: jid.to_bare(),
-                trace: RelayTraceContext::default(),
-            })
-            .await;
-        self.finish_remote_owner_registration_retire(services, jid, registration, detach)
-            .await
-    }
-
-    pub(in super::super) async fn finish_remote_owner_registration_retire(
-        &self,
-        services: &OrderedRelayDeliveryServices,
-        jid: &jid::FullJid,
-        registration: &RemoteOwnerRegistration,
-        detach: Result<RelayForceDetachRemoteUserResourceReply, RelayAskError>,
-    ) -> bool {
-        let reply = match detach {
-            Ok(reply) => reply,
-            Err(error) if ask_error_proves_remote_resource_ref_stale(&error) => {
-                tracing::info!(
-                    jid = %jid,
-                    ?error,
-                    "clustered remote-resource replacement cleaning stale old-socket mirror"
-                );
-                if !unregister_remote_owner_actor_entry(services, jid, &registration.owner).await {
-                    return false;
-                }
-                services
-                    .connection_registry
-                    .unregister_if_owner(jid, &registration.owner);
-                return true;
-            }
-            Err(error) => {
-                tracing::warn!(
-                    jid = %jid,
-                    ?error,
-                    "clustered remote-resource replacement refused uncertain old-socket detach"
-                );
-                return false;
-            }
-        };
-        if !matches!(
-            reply.status,
-            RelayRemoteResourceForceDetachStatus::Detached
-                | RelayRemoteResourceForceDetachStatus::NotLive
-        ) {
-            tracing::warn!(
-                jid = %jid,
-                status = ?reply.status,
-                "clustered remote-resource replacement refused uncertain old-socket detach"
-            );
-            return false;
-        }
-        if !unregister_remote_owner_actor_entry(services, jid, &registration.owner).await {
-            return false;
-        }
-        services
-            .connection_registry
-            .unregister_if_owner(jid, &registration.owner);
-        true
-    }
-
-    pub(in super::super) async fn cleanup_remote_owner_resource_if_registration(
-        &self,
-        jid: &jid::FullJid,
-        registration_id: RemoteResourceRegistrationId,
-    ) {
-        let Some(services) = self.services.get().cloned() else {
-            return;
-        };
-        let registration = self
-            .remote_owner_resources
-            .lock()
-            .await
-            .get(jid)
-            .filter(|registration| registration.registration_id == registration_id)
-            .cloned();
-        let Some(registration) = registration else {
-            return;
-        };
-        let actor_removed = services
-            .user_registry
-            .ask(UnregisterUserResource {
-                jid: jid.clone(),
-                owner: Some(registration.owner.clone()),
-            })
-            .mailbox_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-            .reply_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-            .await
-            .is_ok();
-        if !actor_removed {
-            return;
-        }
-        services
-            .connection_registry
-            .unregister_if_owner(jid, &registration.owner);
-        let mut registrations = self.remote_owner_resources.lock().await;
-        if registrations
-            .get(jid)
-            .is_some_and(|registration| registration.registration_id == registration_id)
-        {
-            registrations.remove(jid);
-        }
-    }
-
     pub(super) async fn lock_for_remote_owner_registration(
         &self,
         jid: &jid::FullJid,
@@ -578,76 +332,6 @@ impl OrderedRelayDeliveryBridge {
             locks.remove(jid);
         }
     }
-}
-
-pub(in super::super) async fn owner_remote_entry_if_current(
-    actor: &ActorRef<waddle_xmpp::registry::user_actor::UserActor>,
-    registry: &ConnectionRegistry,
-    jid: &jid::FullJid,
-    owner: &Arc<AtomicBool>,
-) -> Result<ConnectionEntry, RelayRemoteResourceUpdateStatus> {
-    let actor_entry = match actor
-        .ask(waddle_xmpp::registry::GetConnectionEntry { jid: jid.clone() })
-        .mailbox_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-        .reply_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-        .await
-    {
-        Ok(Some(entry)) => entry,
-        Ok(None) => return Err(RelayRemoteResourceUpdateStatus::StaleRegistration),
-        Err(_) => return Err(RelayRemoteResourceUpdateStatus::Unavailable),
-    };
-    if !Arc::ptr_eq(&actor_entry.carbons_enabled, owner) {
-        return Err(RelayRemoteResourceUpdateStatus::StaleRegistration);
-    }
-    let Some(registry_entry) = registry.entry_if_owner(jid, owner) else {
-        return Err(RelayRemoteResourceUpdateStatus::StaleRegistration);
-    };
-    if !Arc::ptr_eq(
-        &registry_entry.carbons_enabled,
-        &actor_entry.carbons_enabled,
-    ) {
-        return Err(RelayRemoteResourceUpdateStatus::StaleRegistration);
-    }
-    Ok(registry_entry)
-}
-
-pub(super) async fn remote_owner_registration_is_current(
-    services: &OrderedRelayDeliveryServices,
-    jid: &jid::FullJid,
-    registration: &RemoteOwnerRegistration,
-) -> Result<(), RelayRemoteResourceRegistrationStatus> {
-    let actor = match services
-        .user_registry
-        .ask(waddle_xmpp::registry::GetUser {
-            bare_jid: jid.to_bare(),
-        })
-        .mailbox_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-        .reply_timeout(ORDERED_DELIVERY_MAILBOX_TIMEOUT)
-        .await
-    {
-        Ok(Some(actor)) => actor,
-        Ok(None) => return Err(RelayRemoteResourceRegistrationStatus::StaleRegistration),
-        Err(_) => return Err(RelayRemoteResourceRegistrationStatus::Unavailable),
-    };
-    owner_remote_entry_if_current(
-        &actor,
-        &services.connection_registry,
-        jid,
-        &registration.owner,
-    )
-    .await
-    .map(|_| ())
-    .map_err(|status| match status {
-        RelayRemoteResourceUpdateStatus::Updated => {
-            RelayRemoteResourceRegistrationStatus::Registered
-        }
-        RelayRemoteResourceUpdateStatus::StaleRegistration => {
-            RelayRemoteResourceRegistrationStatus::StaleRegistration
-        }
-        RelayRemoteResourceUpdateStatus::Unavailable => {
-            RelayRemoteResourceRegistrationStatus::Unavailable
-        }
-    })
 }
 
 pub(super) async fn unregister_remote_owner_actor_entry(
